Refactor: restructure the tool code
